Core: Move math logic to headers where applicable
