+++ core/math/audio_frame.h
@@ -76,46 +76,46 @@ struct AudioFrame {
 		return levels[p_idx];
 	}
 
-	constexpr AudioFrame operator+(const AudioFrame &p_frame) const { return AudioFrame(left + p_frame.left, right + p_frame.right); }
-	constexpr AudioFrame operator-(const AudioFrame &p_frame) const { return AudioFrame(left - p_frame.left, right - p_frame.right); }
-	constexpr AudioFrame operator*(const AudioFrame &p_frame) const { return AudioFrame(left * p_frame.left, right * p_frame.right); }
-	constexpr AudioFrame operator/(const AudioFrame &p_frame) const { return AudioFrame(left / p_frame.left, right / p_frame.right); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator+(const AudioFrame &p_frame) const { return AudioFrame(left + p_frame.left, right + p_frame.right); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator-(const AudioFrame &p_frame) const { return AudioFrame(left - p_frame.left, right - p_frame.right); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator*(const AudioFrame &p_frame) const { return AudioFrame(left * p_frame.left, right * p_frame.right); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator/(const AudioFrame &p_frame) const { return AudioFrame(left / p_frame.left, right / p_frame.right); }
 
-	constexpr AudioFrame operator+(float p_sample) const { return AudioFrame(left + p_sample, right + p_sample); }
-	constexpr AudioFrame operator-(float p_sample) const { return AudioFrame(left - p_sample, right - p_sample); }
-	constexpr AudioFrame operator*(float p_sample) const { return AudioFrame(left * p_sample, right * p_sample); }
-	constexpr AudioFrame operator/(float p_sample) const { return AudioFrame(left / p_sample, right / p_sample); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator+(float p_sample) const { return AudioFrame(left + p_sample, right + p_sample); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator-(float p_sample) const { return AudioFrame(left - p_sample, right - p_sample); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator*(float p_sample) const { return AudioFrame(left * p_sample, right * p_sample); }
+	_ALWAYS_INLINE_ constexpr AudioFrame operator/(float p_sample) const { return AudioFrame(left / p_sample, right / p_sample); }
 
-	constexpr void operator+=(const AudioFrame &p_frame) {
+	_ALWAYS_INLINE_ constexpr void operator+=(const AudioFrame &p_frame) {
 		left += p_frame.left;
 		right += p_frame.right;
 	}
-	constexpr void operator-=(const AudioFrame &p_frame) {
+	_ALWAYS_INLINE_ constexpr void operator-=(const AudioFrame &p_frame) {
 		left -= p_frame.left;
 		right -= p_frame.right;
 	}
-	constexpr void operator*=(const AudioFrame &p_frame) {
+	_ALWAYS_INLINE_ constexpr void operator*=(const AudioFrame &p_frame) {
 		left *= p_frame.left;
 		right *= p_frame.right;
 	}
-	constexpr void operator/=(const AudioFrame &p_frame) {
+	_ALWAYS_INLINE_ constexpr void operator/=(const AudioFrame &p_frame) {
 		left /= p_frame.left;
 		right /= p_frame.right;
 	}
 
-	constexpr void operator+=(float p_sample) {
+	_ALWAYS_INLINE_ constexpr void operator+=(float p_sample) {
 		left += p_sample;
 		right += p_sample;
 	}
-	constexpr void operator-=(float p_sample) {
+	_ALWAYS_INLINE_ constexpr void operator-=(float p_sample) {
 		left -= p_sample;
 		right -= p_sample;
 	}
-	constexpr void operator*=(float p_sample) {
+	_ALWAYS_INLINE_ constexpr void operator*=(float p_sample) {
 		left *= p_sample;
 		right *= p_sample;
 	}
-	constexpr void operator/=(float p_sample) {
+	_ALWAYS_INLINE_ constexpr void operator/=(float p_sample) {
 		left /= p_sample;
 		right /= p_sample;
 	}
@@ -125,7 +125,7 @@ struct AudioFrame {
 		right = ::undenormalize(right);
 	}
 
-	_FORCE_INLINE_ AudioFrame lerp(const AudioFrame &p_b, float p_t) const {
+	_ALWAYS_INLINE_ constexpr AudioFrame lerp(const AudioFrame &p_b, float p_t) const {
 		AudioFrame res = *this;
 
 		res.left += (p_t * (p_b.left - left));
@@ -135,38 +135,38 @@ struct AudioFrame {
 	}
 
 	// NOLINTBEGIN(cppcoreguidelines-pro-type-member-init)
-	constexpr AudioFrame(float p_left, float p_right) :
+	_ALWAYS_INLINE_ constexpr AudioFrame(float p_left, float p_right) :
 			left(p_left), right(p_right) {}
-	constexpr AudioFrame(const AudioFrame &p_frame) :
+	_ALWAYS_INLINE_ constexpr AudioFrame(const AudioFrame &p_frame) :
 			left(p_frame.left), right(p_frame.right) {}
 	// NOLINTEND(cppcoreguidelines-pro-type-member-init)
 
-	constexpr void operator=(const AudioFrame &p_frame) {
+	_ALWAYS_INLINE_ constexpr void operator=(const AudioFrame &p_frame) {
 		left = p_frame.left;
 		right = p_frame.right;
 	}
 
-	constexpr operator Vector2() const {
+	_ALWAYS_INLINE_ constexpr operator Vector2() const {
 		return Vector2(left, right);
 	}
 
 	// NOLINTBEGIN(cppcoreguidelines-pro-type-member-init)
-	constexpr AudioFrame(const Vector2 &p_v2) :
+	_ALWAYS_INLINE_ constexpr AudioFrame(const Vector2 &p_v2) :
 			left(p_v2.x), right(p_v2.y) {}
-	constexpr AudioFrame() :
+	_ALWAYS_INLINE_ constexpr AudioFrame() :
 			left(0), right(0) {}
 	// NOLINTEND(cppcoreguidelines-pro-type-member-init)
 };
 
-constexpr AudioFrame operator*(float p_scalar, const AudioFrame &p_frame) {
+_ALWAYS_INLINE_ constexpr AudioFrame operator*(float p_scalar, const AudioFrame &p_frame) {
 	return AudioFrame(p_frame.left * p_scalar, p_frame.right * p_scalar);
 }
 
-constexpr AudioFrame operator*(int32_t p_scalar, const AudioFrame &p_frame) {
+_ALWAYS_INLINE_ constexpr AudioFrame operator*(int32_t p_scalar, const AudioFrame &p_frame) {
 	return AudioFrame(p_frame.left * p_scalar, p_frame.right * p_scalar);
 }
 
-constexpr AudioFrame operator*(int64_t p_scalar, const AudioFrame &p_frame) {
+_ALWAYS_INLINE_ constexpr AudioFrame operator*(int64_t p_scalar, const AudioFrame &p_frame) {
 	return AudioFrame(p_frame.left * p_scalar, p_frame.right * p_scalar);
 }
