+++ core/math/color.cpp
@@ -30,84 +30,11 @@
 
 #include "color.h"
 
-#include "color_names.inc"
-#include "core/math/math_funcs.h"
+#include "core/math/color_names.inc"
 #include "core/string/ustring.h"
 #include "core/templates/hash_map.h"
 
 #include "thirdparty/misc/ok_color.h"
-
-uint32_t Color::to_argb32() const {
-	uint32_t c = (uint8_t)Math::round(a * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(r * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(g * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(b * 255.0f);
-
-	return c;
-}
-
-uint32_t Color::to_abgr32() const {
-	uint32_t c = (uint8_t)Math::round(a * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(b * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(g * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(r * 255.0f);
-
-	return c;
-}
-
-uint32_t Color::to_rgba32() const {
-	uint32_t c = (uint8_t)Math::round(r * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(g * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(b * 255.0f);
-	c <<= 8;
-	c |= (uint8_t)Math::round(a * 255.0f);
-
-	return c;
-}
-
-uint64_t Color::to_abgr64() const {
-	uint64_t c = (uint16_t)Math::round(a * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(b * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(g * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(r * 65535.0f);
-
-	return c;
-}
-
-uint64_t Color::to_argb64() const {
-	uint64_t c = (uint16_t)Math::round(a * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(r * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(g * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(b * 65535.0f);
-
-	return c;
-}
-
-uint64_t Color::to_rgba64() const {
-	uint64_t c = (uint16_t)Math::round(r * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(g * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(b * 65535.0f);
-	c <<= 16;
-	c |= (uint16_t)Math::round(a * 65535.0f);
-
-	return c;
-}
 
 void _append_hex(float p_val, char32_t *string) {
 	int v = Math::round(p_val * 255.0f);
@@ -133,106 +60,6 @@ String Color::to_html(bool p_alpha) const {
 	return txt;
 }
 
-float Color::get_h() const {
-	float min = MIN(r, g);
-	min = MIN(min, b);
-	float max = MAX(r, g);
-	max = MAX(max, b);
-
-	float delta = max - min;
-
-	if (delta == 0.0f) {
-		return 0.0f;
-	}
-
-	float h;
-	if (r == max) {
-		h = (g - b) / delta; // between yellow & magenta
-	} else if (g == max) {
-		h = 2 + (b - r) / delta; // between cyan & yellow
-	} else {
-		h = 4 + (r - g) / delta; // between magenta & cyan
-	}
-
-	h /= 6.0f;
-	if (h < 0.0f) {
-		h += 1.0f;
-	}
-
-	return h;
-}
-
-float Color::get_s() const {
-	float min = MIN(r, g);
-	min = MIN(min, b);
-	float max = MAX(r, g);
-	max = MAX(max, b);
-
-	float delta = max - min;
-
-	return (max != 0.0f) ? (delta / max) : 0.0f;
-}
-
-float Color::get_v() const {
-	float max = MAX(r, g);
-	max = MAX(max, b);
-	return max;
-}
-
-void Color::set_hsv(float p_h, float p_s, float p_v, float p_alpha) {
-	int i;
-	float f, p, q, t;
-	a = p_alpha;
-
-	if (p_s == 0.0f) {
-		// Achromatic (gray)
-		r = g = b = p_v;
-		return;
-	}
-
-	p_h *= 6.0f;
-	p_h = Math::fmod(p_h, 6);
-	i = Math::floor(p_h);
-
-	f = p_h - i;
-	p = p_v * (1.0f - p_s);
-	q = p_v * (1.0f - p_s * f);
-	t = p_v * (1.0f - p_s * (1.0f - f));
-
-	switch (i) {
-		case 0: // Red is the dominant color
-			r = p_v;
-			g = t;
-			b = p;
-			break;
-		case 1: // Green is the dominant color
-			r = q;
-			g = p_v;
-			b = p;
-			break;
-		case 2:
-			r = p;
-			g = p_v;
-			b = t;
-			break;
-		case 3: // Blue is the dominant color
-			r = p;
-			g = q;
-			b = p_v;
-			break;
-		case 4:
-			r = t;
-			g = p;
-			b = p_v;
-			break;
-		default: // (5) Red is the dominant color
-			r = p_v;
-			g = p;
-			b = q;
-			break;
-	}
-}
-
 void Color::set_ok_hsl(float p_h, float p_s, float p_l, float p_alpha) {
 	ok_color::HSL hsl;
 	hsl.h = p_h;
@@ -244,52 +71,6 @@ void Color::set_ok_hsl(float p_h, float p_s, float p_l, float p_alpha) {
 	g = c.g;
 	b = c.b;
 	a = c.a;
-}
-
-bool Color::is_equal_approx(const Color &p_color) const {
-	return Math::is_equal_approx(r, p_color.r) && Math::is_equal_approx(g, p_color.g) && Math::is_equal_approx(b, p_color.b) && Math::is_equal_approx(a, p_color.a);
-}
-
-bool Color::is_same(const Color &p_color) const {
-	return Math::is_same(r, p_color.r) && Math::is_same(g, p_color.g) && Math::is_same(b, p_color.b) && Math::is_same(a, p_color.a);
-}
-
-Color Color::clamp(const Color &p_min, const Color &p_max) const {
-	return Color(
-			CLAMP(r, p_min.r, p_max.r),
-			CLAMP(g, p_min.g, p_max.g),
-			CLAMP(b, p_min.b, p_max.b),
-			CLAMP(a, p_min.a, p_max.a));
-}
-
-void Color::invert() {
-	r = 1.0f - r;
-	g = 1.0f - g;
-	b = 1.0f - b;
-}
-
-Color Color::hex(uint32_t p_hex) {
-	float a = (p_hex & 0xFF) / 255.0f;
-	p_hex >>= 8;
-	float b = (p_hex & 0xFF) / 255.0f;
-	p_hex >>= 8;
-	float g = (p_hex & 0xFF) / 255.0f;
-	p_hex >>= 8;
-	float r = (p_hex & 0xFF) / 255.0f;
-
-	return Color(r, g, b, a);
-}
-
-Color Color::hex64(uint64_t p_hex) {
-	float a = (p_hex & 0xFFFF) / 65535.0f;
-	p_hex >>= 16;
-	float b = (p_hex & 0xFFFF) / 65535.0f;
-	p_hex >>= 16;
-	float g = (p_hex & 0xFFFF) / 65535.0f;
-	p_hex >>= 16;
-	float r = (p_hex & 0xFFFF) / 65535.0f;
-
-	return Color(r, g, b, a);
 }
 
 static int _parse_col4(const String &p_str, int p_ofs) {
@@ -307,12 +88,6 @@ static int _parse_col4(const String &p_str, int p_ofs) {
 
 static int _parse_col8(const String &p_str, int p_ofs) {
 	return _parse_col4(p_str, p_ofs) * 16 + _parse_col4(p_str, p_ofs + 1);
-}
-
-Color Color::inverted() const {
-	Color c = *this;
-	c.invert();
-	return c;
 }
 
 Color Color::html(const String &p_rgba) {
@@ -440,30 +215,6 @@ Color Color::from_string(const String &p_string, const Color &p_default) {
 	} else {
 		return named(p_string, p_default);
 	}
-}
-
-Color Color::from_hsv(float p_h, float p_s, float p_v, float p_alpha) {
-	Color c;
-	c.set_hsv(p_h, p_s, p_v, p_alpha);
-	return c;
-}
-
-Color Color::from_rgbe9995(uint32_t p_rgbe) {
-	float r = p_rgbe & 0x1ff;
-	float g = (p_rgbe >> 9) & 0x1ff;
-	float b = (p_rgbe >> 18) & 0x1ff;
-	float e = (p_rgbe >> 27);
-	float m = Math::pow(2.0f, e - 15.0f - 9.0f);
-
-	float rd = r * m;
-	float gd = g * m;
-	float bd = b * m;
-
-	return Color(rd, gd, bd, 1.0f);
-}
-
-Color Color::from_rgba8(int64_t p_r8, int64_t p_g8, int64_t p_b8, int64_t p_a8) {
-	return Color(p_r8 / 255.0f, p_g8 / 255.0f, p_b8 / 255.0f, p_a8 / 255.0f);
 }
 
 Color::operator String() const {
