+++ core/math/quaternion.cpp
@@ -33,10 +33,13 @@
 #include "core/math/basis.h"
 #include "core/string/ustring.h"
 
-real_t Quaternion::angle_to(const Quaternion &p_to) const {
-	real_t d = dot(p_to);
-	// acos does clamping.
-	return Math::acos(d * d * 2 - 1);
+Vector3 Quaternion::xform(const Vector3 &p_v) const {
+#ifdef MATH_CHECKS
+	ERR_FAIL_COND_V_MSG(!is_normalized(), p_v, "The quaternion " + operator String() + " must be normalized.");
+#endif
+	Vector3 u(x, y, z);
+	Vector3 uv = u.cross(p_v);
+	return p_v + ((uv * w) + u.cross(uv)) * ((real_t)2);
 }
 
 Vector3 Quaternion::get_euler(EulerOrder p_order) const {
@@ -46,56 +49,11 @@ Vector3 Quaternion::get_euler(EulerOrder p_order) const {
 	return Basis(*this).get_euler(p_order);
 }
 
-bool Quaternion::is_equal_approx(const Quaternion &p_quaternion) const {
-	return Math::is_equal_approx(x, p_quaternion.x) && Math::is_equal_approx(y, p_quaternion.y) && Math::is_equal_approx(z, p_quaternion.z) && Math::is_equal_approx(w, p_quaternion.w);
-}
-
-bool Quaternion::is_same(const Quaternion &p_quaternion) const {
-	return Math::is_same(x, p_quaternion.x) && Math::is_same(y, p_quaternion.y) && Math::is_same(z, p_quaternion.z) && Math::is_same(w, p_quaternion.w);
-}
-
-bool Quaternion::is_finite() const {
-	return Math::is_finite(x) && Math::is_finite(y) && Math::is_finite(z) && Math::is_finite(w);
-}
-
-real_t Quaternion::length() const {
-	return Math::sqrt(length_squared());
-}
-
-void Quaternion::normalize() {
-	*this /= length();
-}
-
-Quaternion Quaternion::normalized() const {
-	return *this / length();
-}
-
-bool Quaternion::is_normalized() const {
-	return Math::is_equal_approx(length_squared(), 1, (real_t)UNIT_EPSILON); //use less epsilon
-}
-
 Quaternion Quaternion::inverse() const {
 #ifdef MATH_CHECKS
 	ERR_FAIL_COND_V_MSG(!is_normalized(), Quaternion(), "The quaternion " + operator String() + " must be normalized.");
 #endif
 	return Quaternion(-x, -y, -z, w);
-}
-
-Quaternion Quaternion::log() const {
-	Quaternion src = *this;
-	Vector3 src_v = src.get_axis() * src.get_angle();
-	return Quaternion(src_v.x, src_v.y, src_v.z, 0);
-}
-
-Quaternion Quaternion::exp() const {
-	Quaternion src = *this;
-	Vector3 src_v = Vector3(src.x, src.y, src.z);
-	real_t theta = src_v.length();
-	src_v = src_v.normalized();
-	if (theta < CMP_EPSILON || !src_v.is_normalized()) {
-		return Quaternion(0, 0, 0, 1);
-	}
-	return Quaternion(src_v, theta);
 }
 
 Quaternion Quaternion::slerp(const Quaternion &p_to, real_t p_weight) const {
@@ -264,22 +222,6 @@ Quaternion Quaternion::spherical_cubic_interpolate_in_time(const Quaternion &p_b
 	return q1.slerp(q2, p_weight);
 }
 
-Quaternion::operator String() const {
-	return "(" + String::num_real(x, false) + ", " + String::num_real(y, false) + ", " + String::num_real(z, false) + ", " + String::num_real(w, false) + ")";
-}
-
-Vector3 Quaternion::get_axis() const {
-	if (Math::abs(w) > 1 - CMP_EPSILON) {
-		return Vector3(x, y, z);
-	}
-	real_t r = ((real_t)1) / Math::sqrt(1 - w * w);
-	return Vector3(x * r, y * r, z * r);
-}
-
-real_t Quaternion::get_angle() const {
-	return 2 * Math::acos(w);
-}
-
 Quaternion::Quaternion(const Vector3 &p_axis, real_t p_angle) {
 #ifdef MATH_CHECKS
 	ERR_FAIL_COND_MSG(!p_axis.is_normalized(), "The axis Vector3 " + p_axis.operator String() + " must be normalized.");
@@ -301,29 +243,6 @@ Quaternion::Quaternion(const Vector3 &p_axis, real_t p_angle) {
 	}
 }
 
-// Euler constructor expects a vector containing the Euler angles in the format
-// (ax, ay, az), where ax is the angle of rotation around x axis,
-// and similar for other axes.
-// This implementation uses YXZ convention (Z is the first rotation).
-Quaternion Quaternion::from_euler(const Vector3 &p_euler) {
-	real_t half_a1 = p_euler.y * 0.5f;
-	real_t half_a2 = p_euler.x * 0.5f;
-	real_t half_a3 = p_euler.z * 0.5f;
-
-	// R = Y(a1).X(a2).Z(a3) convention for Euler angles.
-	// Conversion to quaternion as listed in https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19770024290.pdf (page A-6)
-	// a3 is the angle of the first rotation, following the notation in this reference.
-
-	real_t cos_a1 = Math::cos(half_a1);
-	real_t sin_a1 = Math::sin(half_a1);
-	real_t cos_a2 = Math::cos(half_a2);
-	real_t sin_a2 = Math::sin(half_a2);
-	real_t cos_a3 = Math::cos(half_a3);
-	real_t sin_a3 = Math::sin(half_a3);
-
-	return Quaternion(
-			sin_a1 * cos_a2 * sin_a3 + cos_a1 * sin_a2 * cos_a3,
-			sin_a1 * cos_a2 * cos_a3 - cos_a1 * sin_a2 * sin_a3,
-			-sin_a1 * sin_a2 * cos_a3 + cos_a1 * cos_a2 * sin_a3,
-			sin_a1 * sin_a2 * sin_a3 + cos_a1 * cos_a2 * cos_a3);
+Quaternion::operator String() const {
+	return "(" + String::num_real(x, false) + ", " + String::num_real(y, false) + ", " + String::num_real(z, false) + ", " + String::num_real(w, false) + ")";
 }
