+++ core/math/quaternion.h
@@ -32,7 +32,6 @@
 
 #include "core/math/math_funcs.h"
 #include "core/math/vector3.h"
-#include "core/string/ustring.h"
 
 struct [[nodiscard]] Quaternion {
 	union {
@@ -54,29 +53,29 @@ struct [[nodiscard]] Quaternion {
 		return components[p_idx];
 	}
 	_FORCE_INLINE_ real_t length_squared() const;
-	bool is_equal_approx(const Quaternion &p_quaternion) const;
-	bool is_same(const Quaternion &p_quaternion) const;
-	bool is_finite() const;
-	real_t length() const;
-	void normalize();
-	Quaternion normalized() const;
-	bool is_normalized() const;
+	_FORCE_INLINE_ bool is_equal_approx(const Quaternion &p_quaternion) const;
+	_FORCE_INLINE_ bool is_same(const Quaternion &p_quaternion) const;
+	_FORCE_INLINE_ bool is_finite() const;
+	_FORCE_INLINE_ real_t length() const;
+	_FORCE_INLINE_ void normalize();
+	_FORCE_INLINE_ Quaternion normalized() const;
+	_FORCE_INLINE_ bool is_normalized() const;
 	Quaternion inverse() const;
-	Quaternion log() const;
-	Quaternion exp() const;
+	_FORCE_INLINE_ Quaternion log() const;
+	_FORCE_INLINE_ Quaternion exp() const;
 	_FORCE_INLINE_ real_t dot(const Quaternion &p_q) const;
-	real_t angle_to(const Quaternion &p_to) const;
+	_FORCE_INLINE_ real_t angle_to(const Quaternion &p_to) const;
 
 	Vector3 get_euler(EulerOrder p_order = EulerOrder::YXZ) const;
-	static Quaternion from_euler(const Vector3 &p_euler);
+	static _FORCE_INLINE_ Quaternion from_euler(const Vector3 &p_euler);
 
 	Quaternion slerp(const Quaternion &p_to, real_t p_weight) const;
 	Quaternion slerpni(const Quaternion &p_to, real_t p_weight) const;
 	Quaternion spherical_cubic_interpolate(const Quaternion &p_b, const Quaternion &p_pre_a, const Quaternion &p_post_b, real_t p_weight) const;
 	Quaternion spherical_cubic_interpolate_in_time(const Quaternion &p_b, const Quaternion &p_pre_a, const Quaternion &p_post_b, real_t p_weight, real_t p_b_t, real_t p_pre_a_t, real_t p_post_b_t) const;
 
-	Vector3 get_axis() const;
-	real_t get_angle() const;
+	_FORCE_INLINE_ Vector3 get_axis() const;
+	_FORCE_INLINE_ real_t get_angle() const;
 
 	_FORCE_INLINE_ void get_axis_angle(Vector3 &r_axis, real_t &r_angle) const {
 		r_angle = 2 * Math::acos(w);
@@ -86,49 +85,41 @@ struct [[nodiscard]] Quaternion {
 		r_axis.z = z * r;
 	}
 
-	constexpr void operator*=(const Quaternion &p_q);
-	constexpr Quaternion operator*(const Quaternion &p_q) const;
+	_FORCE_INLINE_ constexpr void operator*=(const Quaternion &p_q);
+	_FORCE_INLINE_ constexpr Quaternion operator*(const Quaternion &p_q) const;
 
-	_FORCE_INLINE_ Vector3 xform(const Vector3 &p_v) const {
-#ifdef MATH_CHECKS
-		ERR_FAIL_COND_V_MSG(!is_normalized(), p_v, "The quaternion " + operator String() + " must be normalized.");
-#endif
-		Vector3 u(x, y, z);
-		Vector3 uv = u.cross(p_v);
-		return p_v + ((uv * w) + u.cross(uv)) * ((real_t)2);
-	}
-
+	Vector3 xform(const Vector3 &p_v) const;
 	_FORCE_INLINE_ Vector3 xform_inv(const Vector3 &p_v) const {
 		return inverse().xform(p_v);
 	}
 
-	constexpr void operator+=(const Quaternion &p_q);
-	constexpr void operator-=(const Quaternion &p_q);
-	constexpr void operator*=(real_t p_s);
-	constexpr void operator/=(real_t p_s);
-	constexpr Quaternion operator+(const Quaternion &p_q2) const;
-	constexpr Quaternion operator-(const Quaternion &p_q2) const;
-	constexpr Quaternion operator-() const;
-	constexpr Quaternion operator*(real_t p_s) const;
-	constexpr Quaternion operator/(real_t p_s) const;
+	_FORCE_INLINE_ constexpr void operator+=(const Quaternion &p_q);
+	_FORCE_INLINE_ constexpr void operator-=(const Quaternion &p_q);
+	_FORCE_INLINE_ constexpr void operator*=(real_t p_s);
+	_FORCE_INLINE_ constexpr void operator/=(real_t p_s);
+	_FORCE_INLINE_ constexpr Quaternion operator+(const Quaternion &p_q2) const;
+	_FORCE_INLINE_ constexpr Quaternion operator-(const Quaternion &p_q2) const;
+	_FORCE_INLINE_ constexpr Quaternion operator-() const;
+	_FORCE_INLINE_ constexpr Quaternion operator*(real_t p_s) const;
+	_FORCE_INLINE_ constexpr Quaternion operator/(real_t p_s) const;
 
-	constexpr bool operator==(const Quaternion &p_quaternion) const;
-	constexpr bool operator!=(const Quaternion &p_quaternion) const;
+	_FORCE_INLINE_ constexpr bool operator==(const Quaternion &p_quaternion) const;
+	_FORCE_INLINE_ constexpr bool operator!=(const Quaternion &p_quaternion) const;
 
 	operator String() const;
 
-	constexpr Quaternion() :
+	_FORCE_INLINE_ constexpr Quaternion() :
 			x(0), y(0), z(0), w(1) {}
 
-	constexpr Quaternion(real_t p_x, real_t p_y, real_t p_z, real_t p_w) :
+	_FORCE_INLINE_ constexpr Quaternion(real_t p_x, real_t p_y, real_t p_z, real_t p_w) :
 			x(p_x), y(p_y), z(p_z), w(p_w) {}
 
 	Quaternion(const Vector3 &p_axis, real_t p_angle);
 
-	constexpr Quaternion(const Quaternion &p_q) :
+	_FORCE_INLINE_ constexpr Quaternion(const Quaternion &p_q) :
 			x(p_q.x), y(p_q.y), z(p_q.z), w(p_q.w) {}
 
-	constexpr void operator=(const Quaternion &p_q) {
+	_FORCE_INLINE_ constexpr void operator=(const Quaternion &p_q) {
 		x = p_q.x;
 		y = p_q.y;
 		z = p_q.z;
@@ -245,6 +236,96 @@ constexpr Quaternion Quaternion::operator*(const Quaternion &p_q) const {
 	return r;
 }
 
-constexpr Quaternion operator*(real_t p_real, const Quaternion &p_quaternion) {
+_FORCE_INLINE_ constexpr Quaternion operator*(real_t p_real, const Quaternion &p_quaternion) {
 	return p_quaternion * p_real;
+}
+
+real_t Quaternion::angle_to(const Quaternion &p_to) const {
+	real_t d = dot(p_to);
+	// acos does clamping.
+	return Math::acos(d * d * 2 - 1);
+}
+
+bool Quaternion::is_equal_approx(const Quaternion &p_quaternion) const {
+	return Math::is_equal_approx(x, p_quaternion.x) && Math::is_equal_approx(y, p_quaternion.y) && Math::is_equal_approx(z, p_quaternion.z) && Math::is_equal_approx(w, p_quaternion.w);
+}
+
+bool Quaternion::is_same(const Quaternion &p_quaternion) const {
+	return Math::is_same(x, p_quaternion.x) && Math::is_same(y, p_quaternion.y) && Math::is_same(z, p_quaternion.z) && Math::is_same(w, p_quaternion.w);
+}
+
+bool Quaternion::is_finite() const {
+	return Math::is_finite(x) && Math::is_finite(y) && Math::is_finite(z) && Math::is_finite(w);
+}
+
+real_t Quaternion::length() const {
+	return Math::sqrt(length_squared());
+}
+
+void Quaternion::normalize() {
+	*this /= length();
+}
+
+Quaternion Quaternion::normalized() const {
+	return *this / length();
+}
+
+bool Quaternion::is_normalized() const {
+	return Math::is_equal_approx(length_squared(), 1, (real_t)UNIT_EPSILON); //use less epsilon
+}
+
+Quaternion Quaternion::log() const {
+	Quaternion src = *this;
+	Vector3 src_v = src.get_axis() * src.get_angle();
+	return Quaternion(src_v.x, src_v.y, src_v.z, 0);
+}
+
+Quaternion Quaternion::exp() const {
+	Quaternion src = *this;
+	Vector3 src_v = Vector3(src.x, src.y, src.z);
+	real_t theta = src_v.length();
+	src_v = src_v.normalized();
+	if (theta < CMP_EPSILON || !src_v.is_normalized()) {
+		return Quaternion(0, 0, 0, 1);
+	}
+	return Quaternion(src_v, theta);
+}
+
+Vector3 Quaternion::get_axis() const {
+	if (Math::abs(w) > 1 - CMP_EPSILON) {
+		return Vector3(x, y, z);
+	}
+	real_t r = ((real_t)1) / Math::sqrt(1 - w * w);
+	return Vector3(x * r, y * r, z * r);
+}
+
+real_t Quaternion::get_angle() const {
+	return 2 * Math::acos(w);
+}
+
+// Euler constructor expects a vector containing the Euler angles in the format
+// (ax, ay, az), where ax is the angle of rotation around x axis,
+// and similar for other axes.
+// This implementation uses YXZ convention (Z is the first rotation).
+Quaternion Quaternion::from_euler(const Vector3 &p_euler) {
+	real_t half_a1 = p_euler.y * 0.5f;
+	real_t half_a2 = p_euler.x * 0.5f;
+	real_t half_a3 = p_euler.z * 0.5f;
+
+	// R = Y(a1).X(a2).Z(a3) convention for Euler angles.
+	// Conversion to quaternion as listed in https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19770024290.pdf (page A-6)
+	// a3 is the angle of the first rotation, following the notation in this reference.
+
+	real_t cos_a1 = Math::cos(half_a1);
+	real_t sin_a1 = Math::sin(half_a1);
+	real_t cos_a2 = Math::cos(half_a2);
+	real_t sin_a2 = Math::sin(half_a2);
+	real_t cos_a3 = Math::cos(half_a3);
+	real_t sin_a3 = Math::sin(half_a3);
+
+	return Quaternion(
+			sin_a1 * cos_a2 * sin_a3 + cos_a1 * sin_a2 * cos_a3,
+			sin_a1 * cos_a2 * cos_a3 - cos_a1 * sin_a2 * sin_a3,
+			-sin_a1 * sin_a2 * cos_a3 + cos_a1 * cos_a2 * sin_a3,
+			sin_a1 * sin_a2 * sin_a3 + cos_a1 * cos_a2 * cos_a3);
 }
