+++ core/math/rect2.cpp
@@ -34,83 +34,6 @@
 #include "core/math/transform_2d.h"
 #include "core/string/ustring.h"
 
-bool Rect2::is_equal_approx(const Rect2 &p_rect) const {
-	return position.is_equal_approx(p_rect.position) && size.is_equal_approx(p_rect.size);
-}
-
-bool Rect2::is_same(const Rect2 &p_rect) const {
-	return position.is_same(p_rect.position) && size.is_same(p_rect.size);
-}
-
-bool Rect2::is_finite() const {
-	return position.is_finite() && size.is_finite();
-}
-
-bool Rect2::intersects_segment(const Point2 &p_from, const Point2 &p_to, Point2 *r_pos, Point2 *r_normal) const {
-#ifdef MATH_CHECKS
-	if (unlikely(size.x < 0 || size.y < 0)) {
-		ERR_PRINT("Rect2 size is negative, this is not supported. Use Rect2.abs() to get a Rect2 with a positive size.");
-	}
-#endif
-	real_t min = 0, max = 1;
-	int axis = 0;
-	real_t sign = 0;
-
-	for (int i = 0; i < 2; i++) {
-		real_t seg_from = p_from[i];
-		real_t seg_to = p_to[i];
-		real_t box_begin = position[i];
-		real_t box_end = box_begin + size[i];
-		real_t cmin, cmax;
-		real_t csign;
-
-		if (seg_from < seg_to) {
-			if (seg_from > box_end || seg_to < box_begin) {
-				return false;
-			}
-			real_t length = seg_to - seg_from;
-			cmin = (seg_from < box_begin) ? ((box_begin - seg_from) / length) : 0;
-			cmax = (seg_to > box_end) ? ((box_end - seg_from) / length) : 1;
-			csign = -1.0;
-
-		} else {
-			if (seg_to > box_end || seg_from < box_begin) {
-				return false;
-			}
-			real_t length = seg_to - seg_from;
-			cmin = (seg_from > box_end) ? (box_end - seg_from) / length : 0;
-			cmax = (seg_to < box_begin) ? (box_begin - seg_from) / length : 1;
-			csign = 1.0;
-		}
-
-		if (cmin > min) {
-			min = cmin;
-			axis = i;
-			sign = csign;
-		}
-		if (cmax < max) {
-			max = cmax;
-		}
-		if (max < min) {
-			return false;
-		}
-	}
-
-	Vector2 rel = p_to - p_from;
-
-	if (r_normal) {
-		Vector2 normal;
-		normal[axis] = sign;
-		*r_normal = normal;
-	}
-
-	if (r_pos) {
-		*r_pos = p_from + rel * min;
-	}
-
-	return true;
-}
-
 bool Rect2::intersects_transformed(const Transform2D &p_xform, const Rect2 &p_rect) const {
 #ifdef MATH_CHECKS
 	if (unlikely(size.x < 0 || size.y < 0 || p_rect.size.x < 0 || p_rect.size.y < 0)) {
