+++ core/math/transform_2d.h
@@ -35,8 +35,6 @@
 #include "core/math/vector2.h"
 #include "core/templates/vector.h"
 
-class String;
-
 struct [[nodiscard]] Transform2D {
 	// WARNING: The basis of Transform2D is stored differently from Basis.
 	// In terms of columns array, the basis matrix looks like "on paper":
@@ -61,65 +59,65 @@ struct [[nodiscard]] Transform2D {
 	_FORCE_INLINE_ real_t tdotx(const Vector2 &p_v) const { return columns[0][0] * p_v.x + columns[1][0] * p_v.y; }
 	_FORCE_INLINE_ real_t tdoty(const Vector2 &p_v) const { return columns[0][1] * p_v.x + columns[1][1] * p_v.y; }
 
-	constexpr const Vector2 &operator[](int p_idx) const { return columns[p_idx]; }
-	constexpr Vector2 &operator[](int p_idx) { return columns[p_idx]; }
+	_FORCE_INLINE_ constexpr const Vector2 &operator[](int p_idx) const { return columns[p_idx]; }
+	_FORCE_INLINE_ constexpr Vector2 &operator[](int p_idx) { return columns[p_idx]; }
 
-	void invert();
-	Transform2D inverse() const;
+	_FORCE_INLINE_ void invert();
+	_FORCE_INLINE_ Transform2D inverse() const;
 
-	void affine_invert();
-	Transform2D affine_inverse() const;
+	_FORCE_INLINE_ void affine_invert();
+	_FORCE_INLINE_ Transform2D affine_inverse() const;
 
-	void set_rotation(real_t p_rot);
-	real_t get_rotation() const;
-	real_t get_skew() const;
-	void set_skew(real_t p_angle);
+	_FORCE_INLINE_ void set_rotation(real_t p_rot);
+	_FORCE_INLINE_ real_t get_rotation() const;
+	_FORCE_INLINE_ real_t get_skew() const;
+	_FORCE_INLINE_ void set_skew(real_t p_angle);
 	_FORCE_INLINE_ void set_rotation_and_scale(real_t p_rot, const Size2 &p_scale);
 	_FORCE_INLINE_ void set_rotation_scale_and_skew(real_t p_rot, const Size2 &p_scale, real_t p_skew);
-	void rotate(real_t p_angle);
+	_FORCE_INLINE_ void rotate(real_t p_angle);
 
-	void scale(const Size2 &p_scale);
-	void scale_basis(const Size2 &p_scale);
-	void translate_local(real_t p_tx, real_t p_ty);
-	void translate_local(const Vector2 &p_translation);
+	_FORCE_INLINE_ void scale(const Size2 &p_scale);
+	_FORCE_INLINE_ void scale_basis(const Size2 &p_scale);
+	_FORCE_INLINE_ void translate_local(real_t p_tx, real_t p_ty);
+	_FORCE_INLINE_ void translate_local(const Vector2 &p_translation);
 
-	real_t determinant() const;
+	_FORCE_INLINE_ real_t determinant() const;
 
-	Size2 get_scale() const;
-	void set_scale(const Size2 &p_scale);
+	_FORCE_INLINE_ Size2 get_scale() const;
+	_FORCE_INLINE_ void set_scale(const Size2 &p_scale);
 
 	_FORCE_INLINE_ const Vector2 &get_origin() const { return columns[2]; }
 	_FORCE_INLINE_ void set_origin(const Vector2 &p_origin) { columns[2] = p_origin; }
 
-	Transform2D scaled(const Size2 &p_scale) const;
-	Transform2D scaled_local(const Size2 &p_scale) const;
-	Transform2D translated(const Vector2 &p_offset) const;
-	Transform2D translated_local(const Vector2 &p_offset) const;
-	Transform2D rotated(real_t p_angle) const;
-	Transform2D rotated_local(real_t p_angle) const;
+	_FORCE_INLINE_ Transform2D scaled(const Size2 &p_scale) const;
+	_FORCE_INLINE_ Transform2D scaled_local(const Size2 &p_scale) const;
+	_FORCE_INLINE_ Transform2D translated(const Vector2 &p_offset) const;
+	_FORCE_INLINE_ Transform2D translated_local(const Vector2 &p_offset) const;
+	_FORCE_INLINE_ Transform2D rotated(real_t p_angle) const;
+	_FORCE_INLINE_ Transform2D rotated_local(real_t p_angle) const;
 
-	Transform2D untranslated() const;
+	_FORCE_INLINE_ Transform2D untranslated() const;
 
-	void orthonormalize();
-	Transform2D orthonormalized() const;
-	bool is_conformal() const;
-	bool is_equal_approx(const Transform2D &p_transform) const;
-	bool is_same(const Transform2D &p_transform) const;
-	bool is_finite() const;
+	_FORCE_INLINE_ void orthonormalize();
+	_FORCE_INLINE_ Transform2D orthonormalized() const;
+	_FORCE_INLINE_ bool is_conformal() const;
+	_FORCE_INLINE_ bool is_equal_approx(const Transform2D &p_transform) const;
+	_FORCE_INLINE_ bool is_same(const Transform2D &p_transform) const;
+	_FORCE_INLINE_ bool is_finite() const;
 
-	Transform2D looking_at(const Vector2 &p_target) const;
+	_FORCE_INLINE_ Transform2D looking_at(const Vector2 &p_target) const;
 
-	constexpr bool operator==(const Transform2D &p_transform) const;
-	constexpr bool operator!=(const Transform2D &p_transform) const;
+	_FORCE_INLINE_ constexpr bool operator==(const Transform2D &p_transform) const;
+	_FORCE_INLINE_ constexpr bool operator!=(const Transform2D &p_transform) const;
 
-	void operator*=(const Transform2D &p_transform);
-	Transform2D operator*(const Transform2D &p_transform) const;
-	constexpr void operator*=(real_t p_val);
-	constexpr Transform2D operator*(real_t p_val) const;
-	constexpr void operator/=(real_t p_val);
-	constexpr Transform2D operator/(real_t p_val) const;
+	_FORCE_INLINE_ void operator*=(const Transform2D &p_transform);
+	_FORCE_INLINE_ Transform2D operator*(const Transform2D &p_transform) const;
+	_FORCE_INLINE_ constexpr void operator*=(real_t p_val);
+	_FORCE_INLINE_ constexpr Transform2D operator*(real_t p_val) const;
+	_FORCE_INLINE_ constexpr void operator/=(real_t p_val);
+	_FORCE_INLINE_ constexpr Transform2D operator/(real_t p_val) const;
 
-	Transform2D interpolate_with(const Transform2D &p_transform, real_t p_c) const;
+	_FORCE_INLINE_ Transform2D interpolate_with(const Transform2D &p_transform, real_t p_c) const;
 
 	_FORCE_INLINE_ Vector2 basis_xform(const Vector2 &p_vec) const;
 	_FORCE_INLINE_ Vector2 basis_xform_inv(const Vector2 &p_vec) const;
@@ -132,21 +130,21 @@ struct [[nodiscard]] Transform2D {
 
 	operator String() const;
 
-	constexpr Transform2D(real_t p_xx, real_t p_xy, real_t p_yx, real_t p_yy, real_t p_ox, real_t p_oy) :
+	_FORCE_INLINE_ constexpr Transform2D(real_t p_xx, real_t p_xy, real_t p_yx, real_t p_yy, real_t p_ox, real_t p_oy) :
 			columns{
 				{ p_xx, p_xy },
 				{ p_yx, p_yy },
 				{ p_ox, p_oy },
 			} {}
 
-	constexpr Transform2D(const Vector2 &p_x, const Vector2 &p_y, const Vector2 &p_origin) :
+	_FORCE_INLINE_ constexpr Transform2D(const Vector2 &p_x, const Vector2 &p_y, const Vector2 &p_origin) :
 			columns{ p_x, p_y, p_origin } {}
 
-	Transform2D(real_t p_rot, const Vector2 &p_pos);
+	_FORCE_INLINE_ Transform2D(real_t p_rot, const Vector2 &p_pos);
 
-	Transform2D(real_t p_rot, const Size2 &p_scale, real_t p_skew, const Vector2 &p_pos);
+	_FORCE_INLINE_ Transform2D(real_t p_rot, const Size2 &p_scale, real_t p_skew, const Vector2 &p_pos);
 
-	Transform2D() = default;
+	_FORCE_INLINE_ constexpr Transform2D() = default;
 };
 
 constexpr bool Transform2D::operator==(const Transform2D &p_transform) const {
@@ -288,4 +286,239 @@ Vector<Vector2> Transform2D::xform_inv(const Vector<Vector2> &p_array) const {
 		w[i] = xform_inv(r[i]);
 	}
 	return array;
+}
+
+void Transform2D::invert() {
+	// FIXME: this function assumes the basis is a rotation matrix, with no scaling.
+	// Transform2D::affine_inverse can handle matrices with scaling, so GDScript should eventually use that.
+	SWAP(columns[0][1], columns[1][0]);
+	columns[2] = basis_xform(-columns[2]);
+}
+
+Transform2D Transform2D::inverse() const {
+	Transform2D inv = *this;
+	inv.invert();
+	return inv;
+}
+
+void Transform2D::affine_invert() {
+	real_t det = determinant();
+#ifdef MATH_CHECKS
+	ERR_FAIL_COND(det == 0);
+#endif
+	real_t idet = 1.0f / det;
+
+	SWAP(columns[0][0], columns[1][1]);
+	columns[0] *= Vector2(idet, -idet);
+	columns[1] *= Vector2(-idet, idet);
+
+	columns[2] = basis_xform(-columns[2]);
+}
+
+Transform2D Transform2D::affine_inverse() const {
+	Transform2D inv = *this;
+	inv.affine_invert();
+	return inv;
+}
+
+void Transform2D::rotate(real_t p_angle) {
+	*this = Transform2D(p_angle, Vector2()) * (*this);
+}
+
+real_t Transform2D::get_skew() const {
+	real_t det = determinant();
+	return Math::acos(columns[0].normalized().dot(SIGN(det) * columns[1].normalized())) - (real_t)Math::PI * 0.5f;
+}
+
+void Transform2D::set_skew(real_t p_angle) {
+	real_t det = determinant();
+	columns[1] = SIGN(det) * columns[0].rotated(((real_t)Math::PI * 0.5f + p_angle)).normalized() * columns[1].length();
+}
+
+real_t Transform2D::get_rotation() const {
+	return Math::atan2(columns[0].y, columns[0].x);
+}
+
+void Transform2D::set_rotation(real_t p_rot) {
+	Size2 scale = get_scale();
+	real_t cr = Math::cos(p_rot);
+	real_t sr = Math::sin(p_rot);
+	columns[0][0] = cr;
+	columns[0][1] = sr;
+	columns[1][0] = -sr;
+	columns[1][1] = cr;
+	set_scale(scale);
+}
+
+Transform2D::Transform2D(real_t p_rot, const Vector2 &p_pos) {
+	real_t cr = Math::cos(p_rot);
+	real_t sr = Math::sin(p_rot);
+	columns[0][0] = cr;
+	columns[0][1] = sr;
+	columns[1][0] = -sr;
+	columns[1][1] = cr;
+	columns[2] = p_pos;
+}
+
+Transform2D::Transform2D(real_t p_rot, const Size2 &p_scale, real_t p_skew, const Vector2 &p_pos) {
+	columns[0][0] = Math::cos(p_rot) * p_scale.x;
+	columns[1][1] = Math::cos(p_rot + p_skew) * p_scale.y;
+	columns[1][0] = -Math::sin(p_rot + p_skew) * p_scale.y;
+	columns[0][1] = Math::sin(p_rot) * p_scale.x;
+	columns[2] = p_pos;
+}
+
+Size2 Transform2D::get_scale() const {
+	real_t det_sign = SIGN(determinant());
+	return Size2(columns[0].length(), det_sign * columns[1].length());
+}
+
+void Transform2D::set_scale(const Size2 &p_scale) {
+	columns[0].normalize();
+	columns[1].normalize();
+	columns[0] *= p_scale.x;
+	columns[1] *= p_scale.y;
+}
+
+void Transform2D::scale(const Size2 &p_scale) {
+	scale_basis(p_scale);
+	columns[2] *= p_scale;
+}
+
+void Transform2D::scale_basis(const Size2 &p_scale) {
+	columns[0][0] *= p_scale.x;
+	columns[0][1] *= p_scale.y;
+	columns[1][0] *= p_scale.x;
+	columns[1][1] *= p_scale.y;
+}
+
+void Transform2D::translate_local(real_t p_tx, real_t p_ty) {
+	translate_local(Vector2(p_tx, p_ty));
+}
+
+void Transform2D::translate_local(const Vector2 &p_translation) {
+	columns[2] += basis_xform(p_translation);
+}
+
+void Transform2D::orthonormalize() {
+	// Gram-Schmidt Process
+
+	Vector2 x = columns[0];
+	Vector2 y = columns[1];
+
+	x.normalize();
+	y = y - x * x.dot(y);
+	y.normalize();
+
+	columns[0] = x;
+	columns[1] = y;
+}
+
+Transform2D Transform2D::orthonormalized() const {
+	Transform2D ortho = *this;
+	ortho.orthonormalize();
+	return ortho;
+}
+
+bool Transform2D::is_conformal() const {
+	// Non-flipped case.
+	if (Math::is_equal_approx(columns[0][0], columns[1][1]) && Math::is_equal_approx(columns[0][1], -columns[1][0])) {
+		return true;
+	}
+	// Flipped case.
+	if (Math::is_equal_approx(columns[0][0], -columns[1][1]) && Math::is_equal_approx(columns[0][1], columns[1][0])) {
+		return true;
+	}
+	return false;
+}
+
+bool Transform2D::is_equal_approx(const Transform2D &p_transform) const {
+	return columns[0].is_equal_approx(p_transform.columns[0]) && columns[1].is_equal_approx(p_transform.columns[1]) && columns[2].is_equal_approx(p_transform.columns[2]);
+}
+
+bool Transform2D::is_same(const Transform2D &p_transform) const {
+	return columns[0].is_same(p_transform.columns[0]) && columns[1].is_same(p_transform.columns[1]) && columns[2].is_same(p_transform.columns[2]);
+}
+
+bool Transform2D::is_finite() const {
+	return columns[0].is_finite() && columns[1].is_finite() && columns[2].is_finite();
+}
+
+Transform2D Transform2D::looking_at(const Vector2 &p_target) const {
+	Transform2D return_trans = Transform2D(get_rotation(), get_origin());
+	Vector2 target_position = affine_inverse().xform(p_target);
+	return_trans.set_rotation(return_trans.get_rotation() + (target_position * get_scale()).angle());
+	return return_trans;
+}
+
+void Transform2D::operator*=(const Transform2D &p_transform) {
+	columns[2] = xform(p_transform.columns[2]);
+
+	real_t x0, x1, y0, y1;
+
+	x0 = tdotx(p_transform.columns[0]);
+	x1 = tdoty(p_transform.columns[0]);
+	y0 = tdotx(p_transform.columns[1]);
+	y1 = tdoty(p_transform.columns[1]);
+
+	columns[0][0] = x0;
+	columns[0][1] = x1;
+	columns[1][0] = y0;
+	columns[1][1] = y1;
+}
+
+Transform2D Transform2D::operator*(const Transform2D &p_transform) const {
+	Transform2D t = *this;
+	t *= p_transform;
+	return t;
+}
+
+Transform2D Transform2D::scaled(const Size2 &p_scale) const {
+	// Equivalent to left multiplication
+	Transform2D copy = *this;
+	copy.scale(p_scale);
+	return copy;
+}
+
+Transform2D Transform2D::scaled_local(const Size2 &p_scale) const {
+	// Equivalent to right multiplication
+	return Transform2D(columns[0] * p_scale.x, columns[1] * p_scale.y, columns[2]);
+}
+
+Transform2D Transform2D::untranslated() const {
+	Transform2D copy = *this;
+	copy.columns[2] = Vector2();
+	return copy;
+}
+
+Transform2D Transform2D::translated(const Vector2 &p_offset) const {
+	// Equivalent to left multiplication
+	return Transform2D(columns[0], columns[1], columns[2] + p_offset);
+}
+
+Transform2D Transform2D::translated_local(const Vector2 &p_offset) const {
+	// Equivalent to right multiplication
+	return Transform2D(columns[0], columns[1], columns[2] + basis_xform(p_offset));
+}
+
+Transform2D Transform2D::rotated(real_t p_angle) const {
+	// Equivalent to left multiplication
+	return Transform2D(p_angle, Vector2()) * (*this);
+}
+
+Transform2D Transform2D::rotated_local(real_t p_angle) const {
+	// Equivalent to right multiplication
+	return (*this) * Transform2D(p_angle, Vector2()); // Could be optimized, because origin transform can be skipped.
+}
+
+real_t Transform2D::determinant() const {
+	return columns[0].x * columns[1].y - columns[0].y * columns[1].x;
+}
+
+Transform2D Transform2D::interpolate_with(const Transform2D &p_transform, real_t p_weight) const {
+	return Transform2D(
+			Math::lerp_angle(get_rotation(), p_transform.get_rotation(), p_weight),
+			get_scale().lerp(p_transform.get_scale(), p_weight),
+			Math::lerp_angle(get_skew(), p_transform.get_skew(), p_weight),
+			get_origin().lerp(p_transform.get_origin(), p_weight));
 }
