+++ core/math/triangulate.h
@@ -30,8 +30,11 @@
 
 #pragma once
 
-#include "core/math/vector2.h"
-#include "core/templates/vector.h"
+#include "core/math/math_defs.h"
+
+template <typename T>
+class Vector;
+struct Vector2;
 
 /*
 https://www.flipcode.com/archives/Efficient_Polygon_Triangulation.shtml
