+++ core/math/vector2.h
@@ -30,10 +30,8 @@
 
 #pragma once
 
-#include "core/error/error_macros.h"
 #include "core/math/math_funcs.h"
 
-class String;
 struct Vector2i;
 
 struct [[nodiscard]] Vector2 {
@@ -77,43 +75,43 @@ struct [[nodiscard]] Vector2 {
 		return x < y ? Vector2::AXIS_Y : Vector2::AXIS_X;
 	}
 
-	void normalize();
-	Vector2 normalized() const;
-	bool is_normalized() const;
+	_FORCE_INLINE_ void normalize();
+	_FORCE_INLINE_ Vector2 normalized() const;
+	_FORCE_INLINE_ bool is_normalized() const;
 
-	real_t length() const;
-	real_t length_squared() const;
-	Vector2 limit_length(real_t p_len = 1.0) const;
+	_FORCE_INLINE_ real_t length() const;
+	_FORCE_INLINE_ real_t length_squared() const;
+	_FORCE_INLINE_ Vector2 limit_length(real_t p_len = 1.0) const;
 
-	Vector2 min(const Vector2 &p_vector2) const {
+	_FORCE_INLINE_ Vector2 min(const Vector2 &p_vector2) const {
 		return Vector2(MIN(x, p_vector2.x), MIN(y, p_vector2.y));
 	}
 
-	Vector2 minf(real_t p_scalar) const {
+	_FORCE_INLINE_ Vector2 minf(real_t p_scalar) const {
 		return Vector2(MIN(x, p_scalar), MIN(y, p_scalar));
 	}
 
-	Vector2 max(const Vector2 &p_vector2) const {
+	_FORCE_INLINE_ Vector2 max(const Vector2 &p_vector2) const {
 		return Vector2(MAX(x, p_vector2.x), MAX(y, p_vector2.y));
 	}
 
-	Vector2 maxf(real_t p_scalar) const {
+	_FORCE_INLINE_ Vector2 maxf(real_t p_scalar) const {
 		return Vector2(MAX(x, p_scalar), MAX(y, p_scalar));
 	}
 
-	real_t distance_to(const Vector2 &p_vector2) const;
-	real_t distance_squared_to(const Vector2 &p_vector2) const;
-	real_t angle_to(const Vector2 &p_vector2) const;
-	real_t angle_to_point(const Vector2 &p_vector2) const;
+	_FORCE_INLINE_ real_t distance_to(const Vector2 &p_vector2) const;
+	_FORCE_INLINE_ real_t distance_squared_to(const Vector2 &p_vector2) const;
+	_FORCE_INLINE_ real_t angle_to(const Vector2 &p_vector2) const;
+	_FORCE_INLINE_ real_t angle_to_point(const Vector2 &p_vector2) const;
 	_FORCE_INLINE_ Vector2 direction_to(const Vector2 &p_to) const;
 
-	real_t dot(const Vector2 &p_other) const;
-	real_t cross(const Vector2 &p_other) const;
-	Vector2 posmod(real_t p_mod) const;
-	Vector2 posmodv(const Vector2 &p_modv) const;
-	Vector2 project(const Vector2 &p_to) const;
+	_FORCE_INLINE_ real_t dot(const Vector2 &p_other) const;
+	_FORCE_INLINE_ real_t cross(const Vector2 &p_other) const;
+	_FORCE_INLINE_ Vector2 posmod(real_t p_mod) const;
+	_FORCE_INLINE_ Vector2 posmodv(const Vector2 &p_modv) const;
+	_FORCE_INLINE_ Vector2 project(const Vector2 &p_to) const;
 
-	Vector2 plane_project(real_t p_d, const Vector2 &p_vec) const;
+	_FORCE_INLINE_ Vector2 plane_project(real_t p_d, const Vector2 &p_vec) const;
 
 	_FORCE_INLINE_ Vector2 lerp(const Vector2 &p_to, real_t p_weight) const;
 	_FORCE_INLINE_ Vector2 slerp(const Vector2 &p_to, real_t p_weight) const;
@@ -122,73 +120,73 @@ struct [[nodiscard]] Vector2 {
 	_FORCE_INLINE_ Vector2 bezier_interpolate(const Vector2 &p_control_1, const Vector2 &p_control_2, const Vector2 &p_end, real_t p_t) const;
 	_FORCE_INLINE_ Vector2 bezier_derivative(const Vector2 &p_control_1, const Vector2 &p_control_2, const Vector2 &p_end, real_t p_t) const;
 
-	Vector2 move_toward(const Vector2 &p_to, real_t p_delta) const;
+	_FORCE_INLINE_ Vector2 move_toward(const Vector2 &p_to, real_t p_delta) const;
 
 	Vector2 slide(const Vector2 &p_normal) const;
-	Vector2 bounce(const Vector2 &p_normal) const;
+	_FORCE_INLINE_ Vector2 bounce(const Vector2 &p_normal) const;
 	Vector2 reflect(const Vector2 &p_normal) const;
 
-	bool is_equal_approx(const Vector2 &p_v) const;
-	bool is_same(const Vector2 &p_v) const;
-	bool is_zero_approx() const;
-	bool is_finite() const;
+	_FORCE_INLINE_ bool is_equal_approx(const Vector2 &p_v) const;
+	_FORCE_INLINE_ bool is_same(const Vector2 &p_v) const;
+	_FORCE_INLINE_ bool is_zero_approx() const;
+	_FORCE_INLINE_ bool is_finite() const;
 
-	constexpr Vector2 operator+(const Vector2 &p_v) const;
-	constexpr void operator+=(const Vector2 &p_v);
-	constexpr Vector2 operator-(const Vector2 &p_v) const;
-	constexpr void operator-=(const Vector2 &p_v);
-	constexpr Vector2 operator*(const Vector2 &p_v1) const;
+	_FORCE_INLINE_ constexpr Vector2 operator+(const Vector2 &p_v) const;
+	_FORCE_INLINE_ constexpr void operator+=(const Vector2 &p_v);
+	_FORCE_INLINE_ constexpr Vector2 operator-(const Vector2 &p_v) const;
+	_FORCE_INLINE_ constexpr void operator-=(const Vector2 &p_v);
+	_FORCE_INLINE_ constexpr Vector2 operator*(const Vector2 &p_v1) const;
 
-	constexpr Vector2 operator*(real_t p_rvalue) const;
-	constexpr void operator*=(real_t p_rvalue);
-	constexpr void operator*=(const Vector2 &p_rvalue) { *this = *this * p_rvalue; }
+	_FORCE_INLINE_ constexpr Vector2 operator*(real_t p_rvalue) const;
+	_FORCE_INLINE_ constexpr void operator*=(real_t p_rvalue);
+	_FORCE_INLINE_ constexpr void operator*=(const Vector2 &p_rvalue) { *this = *this * p_rvalue; }
 
-	constexpr Vector2 operator/(const Vector2 &p_v1) const;
+	_FORCE_INLINE_ constexpr Vector2 operator/(const Vector2 &p_v1) const;
 
-	constexpr Vector2 operator/(real_t p_rvalue) const;
+	_FORCE_INLINE_ constexpr Vector2 operator/(real_t p_rvalue) const;
 
-	constexpr void operator/=(real_t p_rvalue);
-	constexpr void operator/=(const Vector2 &p_rvalue) { *this = *this / p_rvalue; }
+	_FORCE_INLINE_ constexpr void operator/=(real_t p_rvalue);
+	_FORCE_INLINE_ constexpr void operator/=(const Vector2 &p_rvalue) { *this = *this / p_rvalue; }
 
-	constexpr Vector2 operator-() const;
+	_FORCE_INLINE_ constexpr Vector2 operator-() const;
 
-	constexpr bool operator==(const Vector2 &p_vec2) const;
-	constexpr bool operator!=(const Vector2 &p_vec2) const;
+	_FORCE_INLINE_ constexpr bool operator==(const Vector2 &p_vec2) const;
+	_FORCE_INLINE_ constexpr bool operator!=(const Vector2 &p_vec2) const;
 
-	constexpr bool operator<(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y < p_vec2.y) : (x < p_vec2.x); }
-	constexpr bool operator>(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y > p_vec2.y) : (x > p_vec2.x); }
-	constexpr bool operator<=(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y <= p_vec2.y) : (x < p_vec2.x); }
-	constexpr bool operator>=(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y >= p_vec2.y) : (x > p_vec2.x); }
+	_FORCE_INLINE_ constexpr bool operator<(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y < p_vec2.y) : (x < p_vec2.x); }
+	_FORCE_INLINE_ constexpr bool operator>(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y > p_vec2.y) : (x > p_vec2.x); }
+	_FORCE_INLINE_ constexpr bool operator<=(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y <= p_vec2.y) : (x < p_vec2.x); }
+	_FORCE_INLINE_ constexpr bool operator>=(const Vector2 &p_vec2) const { return x == p_vec2.x ? (y >= p_vec2.y) : (x > p_vec2.x); }
 
-	real_t angle() const;
-	static Vector2 from_angle(real_t p_angle);
+	_FORCE_INLINE_ real_t angle() const;
+	static _FORCE_INLINE_ Vector2 from_angle(real_t p_angle);
 
 	_FORCE_INLINE_ Vector2 abs() const {
 		return Vector2(Math::abs(x), Math::abs(y));
 	}
 
-	Vector2 rotated(real_t p_by) const;
-	Vector2 orthogonal() const {
+	_FORCE_INLINE_ Vector2 rotated(real_t p_by) const;
+	_FORCE_INLINE_ Vector2 orthogonal() const {
 		return Vector2(y, -x);
 	}
 
-	Vector2 sign() const;
-	Vector2 floor() const;
-	Vector2 ceil() const;
-	Vector2 round() const;
-	Vector2 snapped(const Vector2 &p_by) const;
-	Vector2 snappedf(real_t p_by) const;
-	Vector2 clamp(const Vector2 &p_min, const Vector2 &p_max) const;
-	Vector2 clampf(real_t p_min, real_t p_max) const;
-	real_t aspect() const { return width / height; }
+	_FORCE_INLINE_ Vector2 sign() const;
+	_FORCE_INLINE_ Vector2 floor() const;
+	_FORCE_INLINE_ Vector2 ceil() const;
+	_FORCE_INLINE_ Vector2 round() const;
+	_FORCE_INLINE_ Vector2 snapped(const Vector2 &p_by) const;
+	_FORCE_INLINE_ Vector2 snappedf(real_t p_by) const;
+	_FORCE_INLINE_ Vector2 clamp(const Vector2 &p_min, const Vector2 &p_max) const;
+	_FORCE_INLINE_ Vector2 clampf(real_t p_min, real_t p_max) const;
+	_FORCE_INLINE_ real_t aspect() const { return width / height; }
 
 	operator String() const;
 	operator Vector2i() const;
 
 	// NOLINTBEGIN(cppcoreguidelines-pro-type-member-init)
-	constexpr Vector2() :
+	_FORCE_INLINE_ constexpr Vector2() :
 			x(0), y(0) {}
-	constexpr Vector2(real_t p_x, real_t p_y) :
+	_FORCE_INLINE_ constexpr Vector2(real_t p_x, real_t p_y) :
 			x(p_x), y(p_y) {}
 	// NOLINTEND(cppcoreguidelines-pro-type-member-init)
 };
@@ -307,22 +305,180 @@ Vector2 Vector2::direction_to(const Vector2 &p_to) const {
 	return ret;
 }
 
+real_t Vector2::angle() const {
+	return Math::atan2(y, x);
+}
+
+Vector2 Vector2::from_angle(real_t p_angle) {
+	return Vector2(Math::cos(p_angle), Math::sin(p_angle));
+}
+
+real_t Vector2::length() const {
+	return Math::sqrt(x * x + y * y);
+}
+
+real_t Vector2::length_squared() const {
+	return x * x + y * y;
+}
+
+void Vector2::normalize() {
+	real_t l = x * x + y * y;
+	if (l != 0) {
+		l = Math::sqrt(l);
+		x /= l;
+		y /= l;
+	}
+}
+
+Vector2 Vector2::normalized() const {
+	Vector2 v = *this;
+	v.normalize();
+	return v;
+}
+
+bool Vector2::is_normalized() const {
+	// use length_squared() instead of length() to avoid sqrt(), makes it more stringent.
+	return Math::is_equal_approx(length_squared(), 1, (real_t)UNIT_EPSILON);
+}
+
+real_t Vector2::distance_to(const Vector2 &p_vector2) const {
+	return Math::sqrt((x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y));
+}
+
+real_t Vector2::distance_squared_to(const Vector2 &p_vector2) const {
+	return (x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y);
+}
+
+real_t Vector2::angle_to(const Vector2 &p_vector2) const {
+	return Math::atan2(cross(p_vector2), dot(p_vector2));
+}
+
+real_t Vector2::angle_to_point(const Vector2 &p_vector2) const {
+	return (p_vector2 - *this).angle();
+}
+
+real_t Vector2::dot(const Vector2 &p_other) const {
+	return x * p_other.x + y * p_other.y;
+}
+
+real_t Vector2::cross(const Vector2 &p_other) const {
+	return x * p_other.y - y * p_other.x;
+}
+
+Vector2 Vector2::sign() const {
+	return Vector2(SIGN(x), SIGN(y));
+}
+
+Vector2 Vector2::floor() const {
+	return Vector2(Math::floor(x), Math::floor(y));
+}
+
+Vector2 Vector2::ceil() const {
+	return Vector2(Math::ceil(x), Math::ceil(y));
+}
+
+Vector2 Vector2::round() const {
+	return Vector2(Math::round(x), Math::round(y));
+}
+
+Vector2 Vector2::rotated(real_t p_by) const {
+	real_t sine = Math::sin(p_by);
+	real_t cosi = Math::cos(p_by);
+	return Vector2(
+			x * cosi - y * sine,
+			x * sine + y * cosi);
+}
+
+Vector2 Vector2::posmod(real_t p_mod) const {
+	return Vector2(Math::fposmod(x, p_mod), Math::fposmod(y, p_mod));
+}
+
+Vector2 Vector2::posmodv(const Vector2 &p_modv) const {
+	return Vector2(Math::fposmod(x, p_modv.x), Math::fposmod(y, p_modv.y));
+}
+
+Vector2 Vector2::project(const Vector2 &p_to) const {
+	return p_to * (dot(p_to) / p_to.length_squared());
+}
+
+Vector2 Vector2::clamp(const Vector2 &p_min, const Vector2 &p_max) const {
+	return Vector2(
+			CLAMP(x, p_min.x, p_max.x),
+			CLAMP(y, p_min.y, p_max.y));
+}
+
+Vector2 Vector2::clampf(real_t p_min, real_t p_max) const {
+	return Vector2(
+			CLAMP(x, p_min, p_max),
+			CLAMP(y, p_min, p_max));
+}
+
+Vector2 Vector2::snapped(const Vector2 &p_step) const {
+	return Vector2(
+			Math::snapped(x, p_step.x),
+			Math::snapped(y, p_step.y));
+}
+
+Vector2 Vector2::snappedf(real_t p_step) const {
+	return Vector2(
+			Math::snapped(x, p_step),
+			Math::snapped(y, p_step));
+}
+
+Vector2 Vector2::limit_length(real_t p_len) const {
+	const real_t l = length();
+	Vector2 v = *this;
+	if (l > 0 && p_len < l) {
+		v /= l;
+		v *= p_len;
+	}
+
+	return v;
+}
+
+Vector2 Vector2::move_toward(const Vector2 &p_to, real_t p_delta) const {
+	Vector2 v = *this;
+	Vector2 vd = p_to - v;
+	real_t len = vd.length();
+	return len <= p_delta || len < (real_t)CMP_EPSILON ? p_to : v + vd / len * p_delta;
+}
+
+Vector2 Vector2::bounce(const Vector2 &p_normal) const {
+	return -reflect(p_normal);
+}
+
+bool Vector2::is_equal_approx(const Vector2 &p_v) const {
+	return Math::is_equal_approx(x, p_v.x) && Math::is_equal_approx(y, p_v.y);
+}
+
+bool Vector2::is_same(const Vector2 &p_v) const {
+	return Math::is_same(x, p_v.x) && Math::is_same(y, p_v.y);
+}
+
+bool Vector2::is_zero_approx() const {
+	return Math::is_zero_approx(x) && Math::is_zero_approx(y);
+}
+
+bool Vector2::is_finite() const {
+	return Math::is_finite(x) && Math::is_finite(y);
+}
+
 // Multiplication operators required to workaround issues with LLVM using implicit conversion
 // to Vector2i instead for integers where it should not.
 
-constexpr Vector2 operator*(float p_scalar, const Vector2 &p_vec) {
+_FORCE_INLINE_ constexpr Vector2 operator*(float p_scalar, const Vector2 &p_vec) {
 	return p_vec * p_scalar;
 }
 
-constexpr Vector2 operator*(double p_scalar, const Vector2 &p_vec) {
+_FORCE_INLINE_ constexpr Vector2 operator*(double p_scalar, const Vector2 &p_vec) {
 	return p_vec * p_scalar;
 }
 
-constexpr Vector2 operator*(int32_t p_scalar, const Vector2 &p_vec) {
+_FORCE_INLINE_ constexpr Vector2 operator*(int32_t p_scalar, const Vector2 &p_vec) {
 	return p_vec * p_scalar;
 }
 
-constexpr Vector2 operator*(int64_t p_scalar, const Vector2 &p_vec) {
+_FORCE_INLINE_ constexpr Vector2 operator*(int64_t p_scalar, const Vector2 &p_vec) {
 	return p_vec * p_scalar;
 }
