+++ core/math/vector4.cpp
@@ -30,191 +30,8 @@
 
 #include "vector4.h"
 
-#include "core/math/math_funcs.h"
 #include "core/math/vector4i.h"
 #include "core/string/ustring.h"
-
-Vector4::Axis Vector4::min_axis_index() const {
-	uint32_t min_index = 0;
-	real_t min_value = x;
-	for (uint32_t i = 1; i < 4; i++) {
-		if (operator[](i) <= min_value) {
-			min_index = i;
-			min_value = operator[](i);
-		}
-	}
-	return Vector4::Axis(min_index);
-}
-
-Vector4::Axis Vector4::max_axis_index() const {
-	uint32_t max_index = 0;
-	real_t max_value = x;
-	for (uint32_t i = 1; i < 4; i++) {
-		if (operator[](i) > max_value) {
-			max_index = i;
-			max_value = operator[](i);
-		}
-	}
-	return Vector4::Axis(max_index);
-}
-
-bool Vector4::is_equal_approx(const Vector4 &p_vec4) const {
-	return Math::is_equal_approx(x, p_vec4.x) && Math::is_equal_approx(y, p_vec4.y) && Math::is_equal_approx(z, p_vec4.z) && Math::is_equal_approx(w, p_vec4.w);
-}
-
-bool Vector4::is_same(const Vector4 &p_vec4) const {
-	return Math::is_same(x, p_vec4.x) && Math::is_same(y, p_vec4.y) && Math::is_same(z, p_vec4.z) && Math::is_same(w, p_vec4.w);
-}
-
-bool Vector4::is_zero_approx() const {
-	return Math::is_zero_approx(x) && Math::is_zero_approx(y) && Math::is_zero_approx(z) && Math::is_zero_approx(w);
-}
-
-bool Vector4::is_finite() const {
-	return Math::is_finite(x) && Math::is_finite(y) && Math::is_finite(z) && Math::is_finite(w);
-}
-
-real_t Vector4::length() const {
-	return Math::sqrt(length_squared());
-}
-
-void Vector4::normalize() {
-	real_t lengthsq = length_squared();
-	if (lengthsq == 0) {
-		x = y = z = w = 0;
-	} else {
-		real_t length = Math::sqrt(lengthsq);
-		x /= length;
-		y /= length;
-		z /= length;
-		w /= length;
-	}
-}
-
-Vector4 Vector4::normalized() const {
-	Vector4 v = *this;
-	v.normalize();
-	return v;
-}
-
-bool Vector4::is_normalized() const {
-	return Math::is_equal_approx(length_squared(), (real_t)1, (real_t)UNIT_EPSILON);
-}
-
-real_t Vector4::distance_to(const Vector4 &p_to) const {
-	return (p_to - *this).length();
-}
-
-real_t Vector4::distance_squared_to(const Vector4 &p_to) const {
-	return (p_to - *this).length_squared();
-}
-
-Vector4 Vector4::direction_to(const Vector4 &p_to) const {
-	Vector4 ret(p_to.x - x, p_to.y - y, p_to.z - z, p_to.w - w);
-	ret.normalize();
-	return ret;
-}
-
-Vector4 Vector4::abs() const {
-	return Vector4(Math::abs(x), Math::abs(y), Math::abs(z), Math::abs(w));
-}
-
-Vector4 Vector4::sign() const {
-	return Vector4(SIGN(x), SIGN(y), SIGN(z), SIGN(w));
-}
-
-Vector4 Vector4::floor() const {
-	return Vector4(Math::floor(x), Math::floor(y), Math::floor(z), Math::floor(w));
-}
-
-Vector4 Vector4::ceil() const {
-	return Vector4(Math::ceil(x), Math::ceil(y), Math::ceil(z), Math::ceil(w));
-}
-
-Vector4 Vector4::round() const {
-	return Vector4(Math::round(x), Math::round(y), Math::round(z), Math::round(w));
-}
-
-Vector4 Vector4::lerp(const Vector4 &p_to, real_t p_weight) const {
-	Vector4 res = *this;
-	res.x = Math::lerp(res.x, p_to.x, p_weight);
-	res.y = Math::lerp(res.y, p_to.y, p_weight);
-	res.z = Math::lerp(res.z, p_to.z, p_weight);
-	res.w = Math::lerp(res.w, p_to.w, p_weight);
-	return res;
-}
-
-Vector4 Vector4::cubic_interpolate(const Vector4 &p_b, const Vector4 &p_pre_a, const Vector4 &p_post_b, real_t p_weight) const {
-	Vector4 res = *this;
-	res.x = Math::cubic_interpolate(res.x, p_b.x, p_pre_a.x, p_post_b.x, p_weight);
-	res.y = Math::cubic_interpolate(res.y, p_b.y, p_pre_a.y, p_post_b.y, p_weight);
-	res.z = Math::cubic_interpolate(res.z, p_b.z, p_pre_a.z, p_post_b.z, p_weight);
-	res.w = Math::cubic_interpolate(res.w, p_b.w, p_pre_a.w, p_post_b.w, p_weight);
-	return res;
-}
-
-Vector4 Vector4::cubic_interpolate_in_time(const Vector4 &p_b, const Vector4 &p_pre_a, const Vector4 &p_post_b, real_t p_weight, real_t p_b_t, real_t p_pre_a_t, real_t p_post_b_t) const {
-	Vector4 res = *this;
-	res.x = Math::cubic_interpolate_in_time(res.x, p_b.x, p_pre_a.x, p_post_b.x, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
-	res.y = Math::cubic_interpolate_in_time(res.y, p_b.y, p_pre_a.y, p_post_b.y, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
-	res.z = Math::cubic_interpolate_in_time(res.z, p_b.z, p_pre_a.z, p_post_b.z, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
-	res.w = Math::cubic_interpolate_in_time(res.w, p_b.w, p_pre_a.w, p_post_b.w, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
-	return res;
-}
-
-Vector4 Vector4::posmod(real_t p_mod) const {
-	return Vector4(Math::fposmod(x, p_mod), Math::fposmod(y, p_mod), Math::fposmod(z, p_mod), Math::fposmod(w, p_mod));
-}
-
-Vector4 Vector4::posmodv(const Vector4 &p_modv) const {
-	return Vector4(Math::fposmod(x, p_modv.x), Math::fposmod(y, p_modv.y), Math::fposmod(z, p_modv.z), Math::fposmod(w, p_modv.w));
-}
-
-void Vector4::snap(const Vector4 &p_step) {
-	x = Math::snapped(x, p_step.x);
-	y = Math::snapped(y, p_step.y);
-	z = Math::snapped(z, p_step.z);
-	w = Math::snapped(w, p_step.w);
-}
-
-void Vector4::snapf(real_t p_step) {
-	x = Math::snapped(x, p_step);
-	y = Math::snapped(y, p_step);
-	z = Math::snapped(z, p_step);
-	w = Math::snapped(w, p_step);
-}
-
-Vector4 Vector4::snapped(const Vector4 &p_step) const {
-	Vector4 v = *this;
-	v.snap(p_step);
-	return v;
-}
-
-Vector4 Vector4::snappedf(real_t p_step) const {
-	Vector4 v = *this;
-	v.snapf(p_step);
-	return v;
-}
-
-Vector4 Vector4::inverse() const {
-	return Vector4(1.0f / x, 1.0f / y, 1.0f / z, 1.0f / w);
-}
-
-Vector4 Vector4::clamp(const Vector4 &p_min, const Vector4 &p_max) const {
-	return Vector4(
-			CLAMP(x, p_min.x, p_max.x),
-			CLAMP(y, p_min.y, p_max.y),
-			CLAMP(z, p_min.z, p_max.z),
-			CLAMP(w, p_min.w, p_max.w));
-}
-
-Vector4 Vector4::clampf(real_t p_min, real_t p_max) const {
-	return Vector4(
-			CLAMP(x, p_min, p_max),
-			CLAMP(y, p_min, p_max),
-			CLAMP(z, p_min, p_max),
-			CLAMP(w, p_min, p_max));
-}
 
 Vector4::operator String() const {
 	return "(" + String::num_real(x, true) + ", " + String::num_real(y, true) + ", " + String::num_real(z, true) + ", " + String::num_real(w, true) + ")";
